fix a bug in signup using opt
