+++ e_commerse/base/models.py
@@ -6,6 +6,9 @@ from django.utils.timezone import now, timedelta
 
 
 
+
+
+
 class OTP(models.Model):
     phone_number = models.CharField(max_length=15, unique=True)
     otp = models.CharField(max_length=6)
